Fixed Bug w/Force Schedule, Added Short Log Updates
Fixed a bug with the logic to force a schedule run (even if precipitation exceeded maximum).  It seems in my area, OpenWeatherMaps was reporting 2 inches of precipitation where none had occurred.  After disabling the weather checking, I noticed a logic bug that skipped running the schedule even thought I had set it to force.  

Added feature to the dashboard to refresh the latest log entries in near real time so that you don't have to refresh the entire page to see what is happening.